Reference

Functions to extend with a new model

FitPopulations.initialize! — Function

initialize!(model, parameters)

Initalizes the state of the model.

FitPopulations.parameters — Function

parameters(model)

Returns a named tuple.

FitPopulations.logp — Function

logp(data, model, parameters)

Returns $\log P(data|model, parameters)$.

FitPopulations.sample — Function

sample(rng, data, model, parameters; kw...)

Takes already generated data as input and returns a new data point. This function is called in the simulate function. Keyword arguments are passed through the simulate function.

Population model

FitPopulations.PopulationModel — Type

PopulationModel(model; prior = DiagonalNormalPrior(), shared = ())

Wrap a model for estimating population parameters. Shared parameters should be given as a tuple of symbols.

Fitting

FitPopulations.maximize_logp — Function

maximize_logp(data, model, parameters = parameters(model);
              fixed = (;)
              coupled = [],
              optimizer = default_optimizer(model, parameters, fixed),
              lambda_l2 = 0.,
              hessian_ad = Val(:ForwardDiff),
              gradient_ad = Val(:Enzyme),
              evaluate_training = false,
              evaluate_test_data = nothing,
              evaluation_trigger = EventTrigger(),
              evaluation_options = (;),
              callbacks = [],
              verbosity = 1, print_interval = 3,
              return_g! = false,
              )

See also Callback.

Callbacks

FitPopulations.Callback — Type

Callback(trigger, function)

Callback((trigger1, trigger2, ...), function)

See triggers IterationTrigger, TimeTrigger, EventTrigger. For callback functions see LogProgress, Evaluator, CheckPointSaver.

FitPopulations.LogProgress — Type

LogProgress()

FitPopulations.Evaluator — Type

Evaluator(data, model, label = :evaluation, kw...)

Evaluate logp or mc_marginal_logp on data, model and current parameters. Keyword arguments kw are passed to logp or mc_marginal_logp. Results are saved with the given label.

FitPopulations.CheckPointSaver — Type

CheckPointSaver(filename; overwrite = false)

Saves checkpoints as JLD2 files.

Examples

# save every hour and at the end of the simulation.
Callback((TimeTrigger(3600), EventTrigger((:end,))), CheckPointSaver("fit_results.jld2"))
# save at the start and at the end of the simulation.
Callback(EventTrigger((:start, :end)), CheckPointSaver("fit_results.jld2"))

FitPopulations.TimeTrigger — Type

TimeTrigger(Δt)

Triggers every Δt seconds.

FitPopulations.IterationTrigger — Type

IterationTrigger(Δi)

Triggers every Δi iterations.

FitPopulations.EventTrigger — Type

EventTrigger(events = (:start, :start_finetuner, :start_fallback, :iteration_end, :end))

Triggers at given events.

Optimizers

FitPopulations.Optimizer — Type

Optimizer(; optimizer = OptimOptimizer(Optim.LBFGS()), finetuner = nothing, fallback = OptimisersOptimizer())

Default optimizer. finetuner can be another optimizer that is called after the first optimizer finished. The fallback optimizer is called, if the optimizer or finetuner fails.

Can also be constructed as

Optimizer(optimizer; finetuner = nothing, fallback = OptimisersOptimizer(), kw...)

where optimizer can be a symbol (to use NLopt), or an optimiser from Optim or Optimiser. See also NLoptOptimizer, OptimOptimizer, OptimisersOptimizer.

FitPopulations.NLoptOptimizer — Type

NLoptOptimizer(optimizer; options...)

The optimizer is a symbol (e.g. :LD_LBGFS) as specified here. For options, see NLopt options.

FitPopulations.OptimOptimizer — Type

OptimOptimizer(optimizer; options...)

Optimizer can be anything from subtypes.(subtypes(Optim.AbstractOptimizer)). For options, see Optim Options.

FitPopulations.OptimisersOptimizer — Type

OptimisersOptimizer(opt; maxeval = 10^5, maxtime = Inf, min_grad_norm = 1e-8, lower_bounds = -Inf, upper_bounds = Inf)

Optimizer opt can be anything from subtypes(Optimisers.AbstractRule). Optimization stops, when the L2-norm of the gradient falls below min_grad_norm or maxeval or maxtime is reached. See also Optimisers.

FitPopulations.LaplaceEM — Type

LaplaceEM(model, Estep_optimizer = Optimizer(), derivative_threshold = 1e-3, iterations = 10, stopper = () -> false)

Implements the Expectation-Maximization (EM) method with Laplace approximation, as described e.g. in Huys et al. (2012).

Simulation

FitPopulations.simulate — Function

simulate(
    model,
    parameters;
    n_steps,
    stop,
    init,
    tracked,
    rng,
    kw...
)

Returns a named tuple (; data, logp). stop(data, i) is a boolean function that depends on the sequence of simulated data and the iteration counter i. If tracked = true the state of the model is saved for every step in the simulation. Additional keyword arguments kw are passed to the sample function.

FitPopulations.logp_tracked — Function

logp_tracked(data, model, parameters)

Returns a names tuple (; history, logp). In the history the state of the model is saved for every step.

Evaluation

FitPopulations.mc_marginal_logp — Function

mc_marginal_logp(data, model::PopulationModel, params;
                 repetitions = 20, n_samples = 10^4, rng = Random.default_rng())

Estimate the marginal log probability of the data given a model by sampling from the population.

FitPopulations.BIC_int — Function

BIC_int(data, model, params; kw...)

Estimate the Bayesian Information Criterion by sampling from the population. Keyword arguments kw are passed to mc_marginal_logp.

Derivatives

FitPopulations.gradient_logp — Function

gradient_logp(data, model, parameters; ad = Val(:Enzyme))

Compute the gradient of logp.

FitPopulations.hessian_logp — Function

hessian_logp(data, model, parameters; ad = Val(:ForwardDiff))

Compute the hessian of logp.