Reference

Functions to extend with a new model

FitPopulations.initialize! — Function

initialize!(model, parameters)

Initalizes the state of the model.

FitPopulations.parameters — Function

parameters(model)

Returns a named tuple.

FitPopulations.logp — Function

logp(data, model, parameters)

Returns $\log P(data|model, parameters)$.

FitPopulations.sample — Function

sample(rng, data, model, parameters; kw...)

Takes already generated data as input and returns a new data point. This function is called in the simulate function. Keyword arguments are passed through the simulate function.

Population model

FitPopulations.PopulationModel — Type

PopulationModel(model; prior = DiagonalNormalPrior(), shared = ())

Wrap a model for estimating population parameters. Shared parameters should be given as a tuple of symbols.

Fitting

FitPopulations.maximize_logp — Function

maximize_logp(data, model, parameters = ComponentArray(parameters(model));
              fixed = (;)
              coupled = [],
              optimizer = default_optimizer(model, parameters, fixed),
              lambda_l2 = 0.,
              verbosity = 1, print_interval = 10,
              return_g! = false,
              evaluation = (;),
              kw...)

Simulation

FitPopulations.simulate — Function

simulate(
    model,
    parameters;
    n_steps,
    stop,
    init,
    tracked,
    rng,
    kw...
)

Returns a named tuple (; data, logp). stop(data, i) is a boolean function that depends on the sequence of simulated data and the iteration counter i. If tracked = true the state of the model is saved for every step in the simulation. Additional keyword arguments kw are passed to the sample function.

FitPopulations.logp_tracked — Function

logp_tracked(data, model, parameters)

Returns a names tuple (; history, logp). In the history the state of the model is saved for every step.

Evaluation

FitPopulations.mc_marginal_logp — Function

mc_marginal_logp(data, model::PopulationModel, params;
                 repetitions = 20, n_samples = 10^4, rng = Random.default_rng())

Estimate the marginal log probability of the data given a model by sampling from the population.

FitPopulations.BIC_int — Function

BIC_int(data, model, params; kw...)

Estimate the Bayesian Information Criterion by sampling from the population. Keyword arguments kw are passed to mc_marginal_logp.

Derivatives

FitPopulations.gradient_logp — Function

gradient_logp(data, model, parameters)

Compute the gradient of logp.

FitPopulations.hessian_logp — Function

hessian_logp(data, model, parameters; ad = :ForwardDiff)

Compute the hessian of logp.